feat: Rename packages as they should be! (task7-sprint1-model)
