+++ src/main/java/model/entities/Doctor.java
@@ -1,4 +1,4 @@
-package model.entitiess;
+package model.entities;
 
 import org.bson.types.ObjectId;
