+++ src/main/java/model/entities/Report.java
@@ -1,4 +1,4 @@
-package model.entitiess;
+package model.entities;
 
 import org.bson.types.ObjectId;
