+++ src/main/java/model/mongohelpers/MongoAttributes.java
@@ -1,4 +1,4 @@
-package model.mongohelperss;
+package model.mongohelpers;
 
 public class MongoAttributes {
     public final static String IdAttribute = "_id";
